// types/crypto.tsx
// Types for CoinGecko API responses
export interface Coin {
    id: string;
    symbol: string;
    name: string;
    image: string;
    current_price: number;
    market_cap: number;
    market_cap_rank: number;
    fully_diluted_valuation: number | null;
    total_volume: number;
    high_24h: number;
    low_24h: number;
    price_change_24h: number;
    price_change_percentage_24h: number;
    market_cap_change_24h: number;
    market_cap_change_percentage_24h: number;
    circulating_supply: number;
    total_supply: number | null;
    max_supply: number | null;
    ath: number;
    ath_change_percentage: number;
    ath_date: string;
    atl: number;
    atl_change_percentage: number;
    atl_date: string;
    roi: null | {
      times: number;
      currency: string;
      percentage: number;
    };
    last_updated: string;
  }
  
  export interface SearchCoin {
    id: string;
    name: string;
    api_symbol: string;
    symbol: string;
    market_cap_rank: number;
    thumb: string;
    large: string;
  }
  
  export interface SearchResult {
    coins: SearchCoin[];
    exchanges: any[];
    categories: any[];
    nfts: any[];
  }
  
  export type SortDirection = "asc" | "desc";
  export type SortKey = "rank" | "name" | "price" | "change";